updated some types

+++ app/api/[storeId]/categories/route.ts
@@ -5,7 +5,7 @@ import prismadb from '@/lib/prismadb';
  
 export async function POST(
   req: Request,
-  { params }: { params: { storeId: integer } }
+  { params }: { params: { storeId: bigint } }
 ) {
   try {
     const { userId } = auth();
@@ -58,7 +58,7 @@ export async function POST(
 
 export async function GET(
   req: Request,
-  { params }: { params: { storeId: integer } }
+  { params }: { params: { storeId: bigint } }
 ) {
   try {
     if (!params.storeId) {
